convert browserpet > browserbuddy
