+++ src/assets/projects/browserbuddy-extension.gif
