+++ src/theme/index.css
@@ -120,11 +120,11 @@
   /* brown 663a31 */
   /* tan eec39a */
 
-  --theme-browserpet-primary: #fef8dd;
-  --theme-browserpet-secondary: #51f249;
-  --theme-browserpet-bg: #1fb9f3;
-  --theme-browserpet-neutral: #000000;
-  --theme-browserpet-blobglow: #a4e8a0;
+  --theme-browserbuddy-primary: #fef8dd;
+  --theme-browserbuddy-secondary: #51f249;
+  --theme-browserbuddy-bg: #1fb9f3;
+  --theme-browserbuddy-neutral: #000000;
+  --theme-browserbuddy-blobglow: #a4e8a0;
   /* blue 1fb9f3 */
   /* green 51f249 */
   /* white fef8dd */
